invariant refactor in tetwild
