+++ src/wmtk/EdgeSplittingConn.cpp
@@ -54,8 +54,7 @@ bool wmtk::TetMesh::split_edge(const Tuple& loc0, std::vector<Tuple>& new_edges)
     auto rollback_vert_conn = operation_update_connectivity_impl(new_tet_id, new_tet_conn);
 
     Tuple new_loc = tuple_from_vertex(v_id);
-    if (!vertex_invariant(new_loc) || !edge_invariant(new_loc) || !tetrahedron_invariant(new_loc) ||
-        !split_after(new_loc)) {
+    if (!split_after(new_loc)) {
         m_vertex_connectivity[v_id].m_is_removed = true;
         m_vertex_connectivity[v_id].m_conn_tets.clear();
